Fixed problem with BulkEditingTools and Tabs in IE
When GridFieldBulkEditingTools are used on GridField within Form Tab, Internet
Explorer (11) has problems switching to another Form Tab. In this case, clicking
on different tab has no effect, one must at first click somewhere else on the
page, and only after that it's possible to change the tab. While debugging, I
found out that the reason of this behavior is cloning, moving and removing of
the tr with the bulk manager options. Without cloning and removing, and moving
the tr with filter header or sortable header instead of tr with bulk manager
options, switching tabs works fine even in the IE 11.